Support standalone mode
