+++ README.md
@@ -7,3 +7,25 @@
 Symmetrical Transparent Remote Procedure Calls
 
 The core library that powers [ImJoy](https://imjoy.io).
+
+
+# Install Jupyter notebook extension
+
+
+```bash
+git clone https://github.com/imjoy-team/imjoy-rpc.git
+
+cd imjoy-rpc
+
+# install the extension
+jupyter nbextension install nbextension/imjoy-rpc.js
+
+# for development, add --symlink
+# jupyter nbextension install nbextension/imjoy-rpc.js --symlink
+
+# enable the extension
+jupyter nbextension enable imjoy-rpc
+
+# you can also disabled it if you don't need it anymore
+# jupyter nbextension disable imjoy-rpc
+```
