--- javascript/src/connection/socketio-connection.js
@@ -1,5 +1,0 @@
-import io from "socket.io-client";
-
-const socket = io("http://localhost:8080");
-// TODO: support socketio message forwarding
-socket.on("connect", () => {});
