+++ nbextension/README.md
@@ -1,0 +1,8 @@
+ImJoy RPC
+============
+
+
+Description
+-----------
+Supporting running notebook inside ImJoy
+
